refactored form into separate components
